move publisher to separate file
